refactor: graphir dir structure
